refactor: move json_serdes into gravitino.api.types
apache/gravitino#5202
